requirements.txt
requests
python-dotenv
google-generativeai
Pillow
playwright
playwright-stealth
